Refactoring opentracing out of core code, moving context unit tests into their own test
