+++ src/main/java/com/uber/cadence/internal/worker/LocalActivityWorker.java
@@ -22,6 +22,7 @@ import com.uber.cadence.HistoryEvent;
 import com.uber.cadence.MarkerRecordedEventAttributes;
 import com.uber.cadence.PollForActivityTaskResponse;
 import com.uber.cadence.common.RetryOptions;
+import com.uber.cadence.context.ContextPropagator;
 import com.uber.cadence.internal.common.LocalActivityMarkerData;
 import com.uber.cadence.internal.metrics.MetricsTag;
 import com.uber.cadence.internal.metrics.MetricsType;
@@ -38,6 +39,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.function.BiFunction;
 import java.util.function.Consumer;
 import java.util.function.LongSupplier;
+import java.util.stream.Collectors;
 
 public final class LocalActivityWorker implements SuspendableWorker {
 
@@ -271,9 +273,18 @@ public final class LocalActivityWorker implements SuspendableWorker {
   }
 
   private void restoreContext(Map<String, byte[]> context) {
-    options
-        .getContextPropagators()
-        .forEach(
-            propagator -> propagator.setCurrentContext(propagator.deserializeContext(context)));
+    for (ContextPropagator propagator : options.getContextPropagators()) {
+      // Only send the context propagator the fields that belong to them
+      // Change the map from MyPropagator:foo -> bar to foo -> bar
+      Map<String, byte[]> filteredData =
+          context
+              .entrySet()
+              .stream()
+              .filter(e -> e.getKey().startsWith(propagator.getName()))
+              .collect(
+                  Collectors.toMap(
+                      e -> e.getKey().substring(e.getKey().indexOf(":") + 1), Map.Entry::getValue));
+      propagator.setCurrentContext(propagator.deserializeContext(filteredData));
+    }
   }
 }
