+++ src/test/java/com/uber/cadence/internal/testing/WorkflowTestingTest.java
@@ -18,7 +18,6 @@
 package com.uber.cadence.internal.testing;
 
 import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
 import static org.junit.Assert.fail;
 import static org.mockito.Matchers.anyString;
@@ -44,8 +43,6 @@ import com.uber.cadence.client.WorkflowException;
 import com.uber.cadence.client.WorkflowOptions;
 import com.uber.cadence.client.WorkflowStub;
 import com.uber.cadence.client.WorkflowTimedOutException;
-import com.uber.cadence.context.ContextPropagator;
-import com.uber.cadence.context.OpenTracingContextPropagator;
 import com.uber.cadence.internal.common.WorkflowExecutionUtils;
 import com.uber.cadence.testing.SimulatedTimeoutException;
 import com.uber.cadence.testing.TestEnvironmentOptions;
@@ -53,25 +50,13 @@ import com.uber.cadence.testing.TestWorkflowEnvironment;
 import com.uber.cadence.worker.Worker;
 import com.uber.cadence.workflow.ActivityTimeoutException;
 import com.uber.cadence.workflow.Async;
-import com.uber.cadence.workflow.ChildWorkflowOptions;
 import com.uber.cadence.workflow.ChildWorkflowTimedOutException;
 import com.uber.cadence.workflow.Promise;
 import com.uber.cadence.workflow.SignalMethod;
 import com.uber.cadence.workflow.Workflow;
 import com.uber.cadence.workflow.WorkflowMethod;
-import io.opentracing.Scope;
-import io.opentracing.Span;
-import io.opentracing.Tracer;
-import io.opentracing.mock.MockSpan;
-import io.opentracing.mock.MockTracer;
-import io.opentracing.util.GlobalTracer;
-import io.opentracing.util.ThreadLocalScopeManager;
-import java.nio.charset.StandardCharsets;
 import java.time.Duration;
-import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.CancellationException;
@@ -87,7 +72,6 @@ import org.junit.rules.Timeout;
 import org.junit.runner.Description;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.slf4j.MDC;
 
 public class WorkflowTestingTest {
   private static final Logger log = LoggerFactory.getLogger(WorkflowTestingTest.class);
@@ -111,10 +95,7 @@ public class WorkflowTestingTest {
   public void setUp() {
     TestEnvironmentOptions options =
         new TestEnvironmentOptions.Builder()
-            .setFactoryOptions(
-                new Worker.FactoryOptions.Builder()
-                    .setContextPropagators(Collections.singletonList(new TestContextPropagator()))
-                    .build())
+            .setFactoryOptions(new Worker.FactoryOptions.Builder().build())
             .build();
     testEnvironment = TestWorkflowEnvironment.newInstance(options);
   }
@@ -751,288 +732,6 @@ public class WorkflowTestingTest {
       fail("unreacheable");
     } catch (WorkflowException e) {
       assertTrue(e.getCause() instanceof ChildWorkflowTimedOutException);
-    }
-  }
-
-  public static class TestContextPropagator implements ContextPropagator {
-
-    @Override
-    public String getName() {
-      return this.getClass().getName();
-    }
-
-    @Override
-    public Map<String, byte[]> serializeContext(Object context) {
-      String testKey = (String) context;
-      if (testKey != null) {
-        return Collections.singletonMap("test", testKey.getBytes(StandardCharsets.UTF_8));
-      } else {
-        return Collections.emptyMap();
-      }
-    }
-
-    @Override
-    public Object deserializeContext(Map<String, byte[]> context) {
-      if (context.containsKey("test")) {
-        return new String(context.get("test"), StandardCharsets.UTF_8);
-      } else {
-        return null;
-      }
-    }
-
-    @Override
-    public Object getCurrentContext() {
-      return MDC.get("test");
-    }
-
-    @Override
-    public void setCurrentContext(Object context) {
-      MDC.put("test", String.valueOf(context));
-    }
-  }
-
-  public static class ContextPropagationWorkflowImpl implements TestWorkflow {
-
-    @Override
-    public String workflow1(String input) {
-      // The test value should be in the MDC
-      return MDC.get("test");
-    }
-  }
-
-  @Test
-  public void testWorkflowContextPropagation() {
-    Worker worker = testEnvironment.newWorker(TASK_LIST);
-    worker.registerWorkflowImplementationTypes(ContextPropagationWorkflowImpl.class);
-    testEnvironment.start();
-    MDC.put("test", "testing123");
-    WorkflowClient client = testEnvironment.newWorkflowClient();
-    WorkflowOptions options =
-        new WorkflowOptions.Builder()
-            .setContextPropagators(Collections.singletonList(new TestContextPropagator()))
-            .build();
-    TestWorkflow workflow = client.newWorkflowStub(TestWorkflow.class, options);
-    String result = workflow.workflow1("input1");
-    assertEquals("testing123", result);
-  }
-
-  public static class ContextPropagationParentWorkflowImpl implements ParentWorkflow {
-
-    @Override
-    public String workflow(String input) {
-      // Get the MDC value
-      String mdcValue = MDC.get("test");
-
-      // Fire up a child workflow
-      ChildWorkflowOptions options =
-          new ChildWorkflowOptions.Builder()
-              .setContextPropagators(Collections.singletonList(new TestContextPropagator()))
-              .build();
-      ChildWorkflow child = Workflow.newChildWorkflowStub(ChildWorkflow.class, options);
-
-      String result = child.workflow(mdcValue, Workflow.getWorkflowInfo().getWorkflowId());
-      return result;
-    }
-
-    @Override
-    public void signal(String value) {}
-  }
-
-  public static class ContextPropagationChildWorkflowImpl implements ChildWorkflow {
-
-    @Override
-    public String workflow(String input, String parentId) {
-      String mdcValue = MDC.get("test");
-      return input + mdcValue;
-    }
-  }
-
-  @Test
-  public void testChildWorkflowContextPropagation() {
-    Worker worker = testEnvironment.newWorker(TASK_LIST);
-    worker.registerWorkflowImplementationTypes(
-        ContextPropagationParentWorkflowImpl.class, ContextPropagationChildWorkflowImpl.class);
-    testEnvironment.start();
-    MDC.put("test", "testing123");
-    WorkflowClient client = testEnvironment.newWorkflowClient();
-    WorkflowOptions options =
-        new WorkflowOptions.Builder()
-            .setContextPropagators(Collections.singletonList(new TestContextPropagator()))
-            .build();
-    ParentWorkflow workflow = client.newWorkflowStub(ParentWorkflow.class, options);
-    String result = workflow.workflow("input1");
-    assertEquals("testing123testing123", result);
-  }
-
-  public static class ContextPropagationThreadWorkflowImpl implements TestWorkflow {
-
-    @Override
-    public String workflow1(String input) {
-      Promise<String> asyncPromise = Async.function(this::async);
-      return asyncPromise.get();
-    }
-
-    private String async() {
-      return "async" + MDC.get("test");
-    }
-  }
-
-  @Test
-  public void testThreadContextPropagation() {
-    Worker worker = testEnvironment.newWorker(TASK_LIST);
-    worker.registerWorkflowImplementationTypes(ContextPropagationThreadWorkflowImpl.class);
-    testEnvironment.start();
-    MDC.put("test", "testing123");
-    WorkflowClient client = testEnvironment.newWorkflowClient();
-    WorkflowOptions options =
-        new WorkflowOptions.Builder()
-            .setContextPropagators(Collections.singletonList(new TestContextPropagator()))
-            .build();
-    TestWorkflow workflow = client.newWorkflowStub(TestWorkflow.class, options);
-    String result = workflow.workflow1("input1");
-    assertEquals("asynctesting123", result);
-  }
-
-  public static class ContextActivityImpl implements TestActivity {
-    @Override
-    public String activity1(String input) {
-      return "activity" + MDC.get("test");
-    }
-  }
-
-  public static class ContextPropagationActivityWorkflowImpl implements TestWorkflow {
-    @Override
-    public String workflow1(String input) {
-      ActivityOptions options =
-          new ActivityOptions.Builder()
-              .setScheduleToCloseTimeout(Duration.ofSeconds(5))
-              .setContextPropagators(Collections.singletonList(new TestContextPropagator()))
-              .build();
-      TestActivity activity = Workflow.newActivityStub(TestActivity.class, options);
-      return activity.activity1("foo");
-    }
-  }
-
-  @Test
-  public void testActivityContextPropagation() {
-    Worker worker = testEnvironment.newWorker(TASK_LIST);
-    worker.registerWorkflowImplementationTypes(ContextPropagationActivityWorkflowImpl.class);
-    worker.registerActivitiesImplementations(new ContextActivityImpl());
-    testEnvironment.start();
-    MDC.put("test", "testing123");
-    WorkflowClient client = testEnvironment.newWorkflowClient();
-    WorkflowOptions options =
-        new WorkflowOptions.Builder()
-            .setContextPropagators(Collections.singletonList(new TestContextPropagator()))
-            .build();
-    TestWorkflow workflow = client.newWorkflowStub(TestWorkflow.class, options);
-    String result = workflow.workflow1("input1");
-    assertEquals("activitytesting123", result);
-  }
-
-  public static class DefaultContextPropagationActivityWorkflowImpl implements TestWorkflow {
-    @Override
-    public String workflow1(String input) {
-      ActivityOptions options =
-          new ActivityOptions.Builder().setScheduleToCloseTimeout(Duration.ofSeconds(5)).build();
-      TestActivity activity = Workflow.newActivityStub(TestActivity.class, options);
-      return activity.activity1("foo");
-    }
-  }
-
-  @Test
-  public void testDefaultActivityContextPropagation() {
-    Worker worker = testEnvironment.newWorker(TASK_LIST);
-    worker.registerWorkflowImplementationTypes(DefaultContextPropagationActivityWorkflowImpl.class);
-    worker.registerActivitiesImplementations(new ContextActivityImpl());
-    testEnvironment.start();
-    MDC.put("test", "testing123");
-    WorkflowClient client = testEnvironment.newWorkflowClient();
-    WorkflowOptions options =
-        new WorkflowOptions.Builder()
-            .setContextPropagators(Collections.singletonList(new TestContextPropagator()))
-            .build();
-    TestWorkflow workflow = client.newWorkflowStub(TestWorkflow.class, options);
-    String result = workflow.workflow1("input1");
-    assertEquals("activitytesting123", result);
-  }
-
-  public static class DefaultContextPropagationParentWorkflowImpl implements ParentWorkflow {
-
-    @Override
-    public String workflow(String input) {
-      // Get the MDC value
-      String mdcValue = MDC.get("test");
-
-      // Fire up a child workflow
-      ChildWorkflowOptions options = new ChildWorkflowOptions.Builder().build();
-      ChildWorkflow child = Workflow.newChildWorkflowStub(ChildWorkflow.class, options);
-
-      String result = child.workflow(mdcValue, Workflow.getWorkflowInfo().getWorkflowId());
-      return result;
-    }
-
-    @Override
-    public void signal(String value) {}
-  }
-
-  @Test
-  public void testDefaultChildWorkflowContextPropagation() {
-    Worker worker = testEnvironment.newWorker(TASK_LIST);
-    worker.registerWorkflowImplementationTypes(
-        DefaultContextPropagationParentWorkflowImpl.class,
-        ContextPropagationChildWorkflowImpl.class);
-    testEnvironment.start();
-    MDC.put("test", "testing123");
-    WorkflowClient client = testEnvironment.newWorkflowClient();
-    WorkflowOptions options =
-        new WorkflowOptions.Builder()
-            .setContextPropagators(Collections.singletonList(new TestContextPropagator()))
-            .build();
-    ParentWorkflow workflow = client.newWorkflowStub(ParentWorkflow.class, options);
-    String result = workflow.workflow("input1");
-    assertEquals("testing123testing123", result);
-  }
-
-  public static class OpenTracingContextPropagationWorkflowImpl implements TestWorkflow {
-    @Override
-    public String workflow1(String input) {
-      Tracer tracer = GlobalTracer.get();
-      Span activeSpan = tracer.scopeManager().activeSpan();
-      MockSpan mockSpan = (MockSpan) activeSpan;
-      assertNotNull(activeSpan);
-      assertEquals("TestWorkflow::workflow1", mockSpan.tags().get("resource.name"));
-      return activeSpan.getBaggageItem("foo");
-    }
-  }
-
-  @Test
-  public void testOpenTracingContextPropagation() {
-    MockTracer tracer =
-        new MockTracer(new ThreadLocalScopeManager(), MockTracer.Propagator.TEXT_MAP);
-    GlobalTracer.registerIfAbsent(tracer);
-    Span span = tracer.buildSpan("testContextPropagation").start();
-
-    Worker worker = testEnvironment.newWorker(TASK_LIST);
-    worker.registerWorkflowImplementationTypes(OpenTracingContextPropagationWorkflowImpl.class);
-    testEnvironment.start();
-    WorkflowClient client = testEnvironment.newWorkflowClient();
-    WorkflowOptions options =
-        new WorkflowOptions.Builder()
-            .setContextPropagators(
-                Arrays.asList(new TestContextPropagator(), new OpenTracingContextPropagator()))
-            .build();
-
-    try (Scope scope = tracer.scopeManager().activate(span)) {
-
-      Span activeSpan = tracer.scopeManager().activeSpan();
-      activeSpan.setBaggageItem("foo", "bar");
-
-      TestWorkflow workflow = client.newWorkflowStub(TestWorkflow.class, options);
-      assertEquals("bar", workflow.workflow1("input1"));
-
-    } finally {
-      span.finish();
     }
   }
 }
