+++ src/test/resources/logback-test.xml
@@ -25,6 +25,7 @@
         </encoder>
     </appender>
     <logger name="io.netty" level="INFO"/>
+    <logger name="com.uber.cadence.context" level="DEBUG"/>
     <root level="INFO">
         <appender-ref ref="STDOUT" />
     </root>
